Fixed broken type for designer field on IDR

+++ coresight/src/access_ports/generic_ap/mod.rs
@@ -38,7 +38,7 @@ define_ap!(GenericAP);
 
 define_ap_register!(GenericAP, IDR, 0x0FC, [
         (REVISION: u8),
-        (DESIGNER: u8),
+        (DESIGNER: u16),
         (CLASS: APClass),
         (_RES0: u8),
         (VARIANT: u8),
@@ -47,7 +47,7 @@ define_ap_register!(GenericAP, IDR, 0x0FC, [
     value,
     IDR {
         REVISION: ((value >> 28) & 0x0F) as u8,
-        DESIGNER: ((value >> 17) & 0x7FF) as u8,
+        DESIGNER: ((value >> 17) & 0x7FF) as u16,
         CLASS:    APClass::from_u8(((value >> 13) & 0x0F) as u8).unwrap(),
         _RES0:     0,
         VARIANT:  ((value >> 4) & 0x0F) as u8,
